Moved Type Mapping functionality to linq2db sources.
