+++ Source/LinqToDB/Expressions/WrapperAttribute.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Tests.Playground.TypeMapping
+namespace LinqToDB.Expressions
 {
 	[AttributeUsage(AttributeTargets.Class | AttributeTargets.Struct | AttributeTargets.Interface | AttributeTargets.Enum)]
 	public class WrapperAttribute : Attribute
